+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - Complete rewrite of extension, now being a real editor extension instead of just a collection of snippets
 - Shebang suggestions now only shown at beginning of 1st line and not anywhere in document
 - After completing a shebang line, document language is automatically updated appropritately
-- Support for zsh shebang
+- Support for zsh and Nushell shebangs
 - Shebangs with `/usr/bin/...` and `/usr/bin/env ...` are both supported
 - The extension learns from it's usage, frequently used shebangs are preferred in suggestion list